Move send email notification code into celery task
