+++ requirements.txt
@@ -17,6 +17,7 @@ djangorestframework==3.10.3
 gevent==20.4.0
 gunicorn==20.0.4
 hawkrest==1.0.1
+notifications-python-client==5.5.1
 psycopg2-binary==2.7.5
 raven==6.9.0
 redis==3.4.1
